add a new way of handle imports
